+++ .github/CHANGELOG.md
@@ -28,17 +28,17 @@
   on simulators and hardware.
   [#432](https://github.com/XanaduAI/pennylane/pull/432)
 
-* Added the `QNodeCluster` container class, that allows independent
+* Added the `QNodeCollection` container class, that allows independent
   QNodes to be stored and evaluated simultaneously.
   [#466](https://github.com/XanaduAI/pennylane/pull/466)
 
 * Added high level `qml.sum`, `qml.dot`, `qml.apply` functions
-  that act on QNode clusters.
+  that act on QNode collections.
   [#466](https://github.com/XanaduAI/pennylane/pull/466)
 
 * Added a high level `qml.map` function, that maps a quantum
   circuit template over a list of observables or devices, returning
-  a `QNodeCluster`.
+  a `QNodeCollection`.
   [#466](https://github.com/XanaduAI/pennylane/pull/466)
 
 ### Breaking changes